+++ examples/udp_sender/main.cpp
@@ -26,7 +26,7 @@ protected:
         mSendTimer.start(1000, true, [&](Timer*) {
 
             const char* msg = "hello";
-            uint32_t size = (uint32_t)(strlen(msg) + 1);
+            size_t size = strlen(msg) + 1;
 
             std::cout << "send: " << msg <<
                 " to " << mUdpSender->getReceiverEndPoint().toString() <<
